requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
kafka-python==1.4.7
# psycopg2==2.8.4
PyYAML==5.1.2
requests==2.22.0
SQLAlchemy==1.3.11
urllib3==1.25.7
